fluxo de 100m rasos separado dos demais
