+++ modulo6/case-estante-virtual/src/business/CompetitionBusiness.ts
@@ -42,72 +42,4 @@ export default class CompetitionBusiness {
         }
     }
 
-    insertMeterDash = async (input: meterDashDTO) => {
-        try {
-            if (!input.athlete || !input.time || !input.competition_id) {
-                throw new CustomError(422, "Missin input")
-            }
-           
-            const active = await this.competitionDatabase.verifyCompetitionActive(input.competition_id)
-            
-            if (!active || active === 0) {
-                throw new CustomError(400, "A competição já foi encerrada")
-            }
-
-            const id: string = this.idGenerator.generate()
-
-            const dashInput = {
-                id: id,
-                athlete: input.athlete,
-                time: input.time,
-                competition_id: input.competition_id
-            }
-
-            const newDash = MeterDash.toMeterDashModel(dashInput)
-
-            await this.competitionDatabase.insertMeterDash(newDash)
-
-            const message = "Atleta e tempo adicionados com sucesso!"
-
-            return message
-
-        } catch (error: any) {
-            throw new CustomError(error.statusCode, error.message)
-        }
-    }
-
-    // insertJavelinThrow = async (input: JavelinDTO) => {
-    //     try {
-    //         if (!input.athlete || !input.distance || !input.competition_id) {
-    //             throw new CustomError(422, "Missin input")
-    //         }
-           
-    //         const active = await this.competitionDatabase.verifyCompetitionActive(input.competition_id)
-            
-    //         if (!active || active === 0) {
-    //             throw new CustomError(400, "A competição já foi encerrada")
-    //         }
-
-    //         const id: string = this.idGenerator.generate()
-
-    //         const dashInput = {
-    //             id: id,
-    //             athlete: input.athlete,
-    //             distance: input.distance,
-    //             competition_id: input.competition_id
-    //         }
-
-    //         const newDash = MeterDash.toMetreDashModel(dashInput)
-
-    //         await this.competitionDatabase.insertMeterDash(newDash)
-
-    //         const message = "Atleta e tempo adicionados com sucesso!"
-
-    //         return message
-
-    //     } catch (error: any) {
-    //         throw new CustomError(error.statusCode, error.message)
-    //     }
-    // }
-
 }
