+++ modulo6/case-estante-virtual/src/business/CompetitionRepository.ts
@@ -4,5 +4,4 @@ import MeterDash from "../model/MeterDash";
 export default interface CompetitionRepository{
     createCompetition (competition: Competition): Promise<Competition>
     verifyCompetitionActive (id: string): Promise<any>
-    insertMeterDash (input: MeterDash): Promise<MeterDash> 
 }
