+++ modulo6/case-estante-virtual/src/business/MeterDash/MeterDashRepository.ts
@@ -1,0 +1,6 @@
+import MeterDash from "../../model/MeterDash"
+
+export default interface MeterDashRepository{
+    verifyCompetitionActive (id: string): Promise<any> 
+    insertMeterDash (input: MeterDash): Promise<MeterDash>
+}
