+++ modulo6/case-estante-virtual/src/data/CompetitionDatabase.ts
@@ -39,21 +39,5 @@ export default class CompetitionDatabase extends BaseDatabase implements Competi
         }
     }
 
-    public insertMeterDash = async (input: MeterDash): Promise<MeterDash> => {
-        try {
-            await this.getConnection()
-                .insert({
-                    id: input.getId(),
-                    athlete: input.getAthlete(),
-                    time: input.getTime(),
-                    competition_id: input.getCompetition_id()
-                })
-                .into("Case1_Dash");
 
-            return input
-
-        } catch (error: any) {
-            throw new Error(error.sqlMessage || error.message);
-        }
-    }
 }
