+++ modulo6/case-estante-virtual/src/routes/CompetitionRouter.ts
@@ -6,5 +6,4 @@ export const competitionRouter = express.Router();
 const competitionController = new CompetitionController();
 
 competitionRouter.post("/create", competitionController.createComptetition)
-competitionRouter.post("/meter/:id", competitionController.insertMeterDash)
 // userRouter.post("/login", userController.login);
